+++ creme/compat/__init__.py
@@ -1,5 +1,6 @@
 from .sklearn import convert_creme_to_sklearn
 from .sklearn import convert_sklearn_to_creme
+from .sklearn import CremeClassifierWrapper
 from .sklearn import CremeRegressorWrapper
 from .sklearn import SKLRegressorWrapper
 from .sklearn import SKLClassifierWrapper