+++ creme/optim/lr_schedule.py
@@ -1,13 +1,17 @@
 import abc
 
 
-__all__ = ['ConstantLR', 'InverseScalingLR']
+__all__ = [
+    'ConstantLR',
+    'InverseScalingLR',
+    'OptimalLR'
+]
 
 
 class LRScheduler(abc.ABC):
 
     @abc.abstractmethod
-    def get(self, iteration: int) -> float:
+    def get(self, t: int) -> float:
         pass
 
 
@@ -16,7 +20,7 @@ class ConstantLR(LRScheduler):
     def __init__(self, learning_rate):
         self.learning_rate = learning_rate
 
-    def get(self, iteration):
+    def get(self, t):
         return self.learning_rate
 
 
@@ -26,5 +30,22 @@ class InverseScalingLR(LRScheduler):
         self.learning_rate = learning_rate
         self.power = power
 
-    def get(self, iteration):
-        return self.learning_rate / (iteration + 1) ** self.power
+    def get(self, t):
+        return self.learning_rate / (t + 1) ** self.power
+
+
+class OptimalLR(LRScheduler):
+    """
+
+    References:
+
+    1. `Stochastic Gradient Descent <https://leon.bottou.org/projects/sgd>`_
+
+    """
+
+    def __init__(self, t0=1000, alpha=1e-4):
+        self.t0 = t0
+        self.alpha = alpha
+
+    def get(self, t):
+        return 1. / (self.alpha * (t + self.t0))